content/conduct: add CoC verbatim
Added a page `/conduct` with our Code of Conduct verbatim
instead of just as a link to the specs.
The website serves the entry point and representation of the
project to both technical and non-technical audiences.

A specialized endpoint makes linking easy but also direct
and shows the project's committment to a code of conduct.

Fixes #436