guides/integrations/redis: use updated Jedis integration
The Jedis integration is now available at
https://mvnrepository.com/artifact/io.orijtech.integrations/ocjedis

and it is published under my domain because:
a) publishing under io.opencensus.integrations is a complex issue
that requires multi-level permissions and approvals but also
b) means that the OpenCensus Java team is obligated to maintain
the integration forever

Being able to update artifacts makes developing integrations a lot
smoother but also releases the small team from obligations.
The code was written and released by Orijtech, Inc. and provides
an Apache License on behalf of OpenCensus Authors.

Fixes #474