integrations/sql: retouch ocjdbc to ensure working end-to-end
Retouch the OCJDBC integration guide now that it was successfully
published on Sonatype. This retouch ensures that the end-to-end
example fully compiles. The last edits of the actual integration
introduced an unexpected compulsory option which rendered the
previous guide uncompilable.

This change also involved running `./gradlew goJF` to format
the Java source code.

Updates #527