guides/google_cloud_spanner: update dependencies + ensure compilation
Reported offline by Alex Amies, the Google Cloud Spanner integration
guide was outdated but also couldn't compile.

This change modernizes the guide to use OpenCensus-Java 0.18.0
(the lastest) up from 0.11.0 and the Cloud Spanner Java
dependency to 1.5.0, up from 0.33.0-beta

Also ensuring that the guides run/compile successfully/end-to-end.

This change also ran `./gradlew goJF` on the Java source code
as well as followed Alex's recommendation of removing usages
of "this.*" except when dealing with scoping problems, as that style
is the norm in Java.

Supersedes #540